+++ packages/cashscript/src/index.ts
@@ -21,4 +21,4 @@ export {
   MockNetworkProvider,
 } from './network/index.js';
 export { randomUtxo, randomToken, randomNFT } from './utils.js';
-export * from './wc-utils.js';
+export * from './walletconnect-utils.js';